% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_heat_stress.R
\name{cal_heat_stress}
\alias{cal_heat_stress}
\title{cal_heat_stress}
\usage{
cal_heat_stress(TempRes, SECTOR, HS, YEAR_INPUT, ...)
}
\arguments{
\item{TempRes}{temporal resolution of input, "day" or "month"}

\item{SECTOR}{choice of sector}

\item{HS}{choice of heat stress function}

\item{YEAR_INPUT}{a vector of years of interest}

\item{...}{individual .nc file name of climate variables in order}
}
\value{
A raster stack of daily grid-level heat stress level
}
\description{
Calculate Heat Stress Levels Based on Climate Projections
}
\details{
The \code{cal_heat_stress} function computes daily or monthly grid-level heat stress levels for specified
sectors, using a user-defined heat stress function and climate data. The function
supports both daily and monthly climate inputs and evaluates heat stress over a specific year.
}
\examples{
\dontrun{
GD_HS <- cal_heat_stress(
  TempRes = "day", SECTOR = "MAIZ_I", HS = WBGT1, YEAR_INPUT = 2027,
  "hurs_day_GFDL-ESM2M_rcp60_r1i1p1_EWEMBI_20210101-20301231.nc4",
  "tas_day_GFDL-ESM2M_rcp60_r1i1p1_EWEMBI_20210101-20301231.nc4",
  "ps_day_GFDL-ESM2M_rcp60_r1i1p1_EWEMBI_20210101-20301231.nc4"
)
}
}
