% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_pwc.R
\name{cal_pwc}
\alias{cal_pwc}
\title{cal_pwc}
\usage{
cal_pwc(WBGT, LHR, workload)
}
\arguments{
\item{WBGT}{output from HeatStress function}

\item{LHR}{choice of labor heat response function}

\item{workload}{choice of workload intensity, 'high', 'moderate', and, 'low'}
}
\value{
A raster stack of daily grid-level physical work capacity multiplier: ranges from 0 to 1
}
\description{
Calculate Physical Work Capacity (PWC)
}
\details{
The \code{cal_pwc} function computes daily grid-level physical work capacity (PWC),
which range from 0 to 1, based on heat stress levels, a chosen labor heat response function,
and workload intensity.
}
\examples{
\dontrun{
GD_PWC <- cal_pwc(HeatStress = HS.stack.y, LHR = LHR_Foster, workload = "high")
}
}
