% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_to_monthly.R
\name{daily_to_monthly}
\alias{daily_to_monthly}
\title{daily_to_monthly}
\usage{
daily_to_monthly(input_rack)
}
\arguments{
\item{input_rack}{Raster rack of daily value}
}
\value{
A raster rack of monthly value
}
\description{
Convert Daily Raster Data to Monthly Averages
}
\details{
This function takes a raster stack of daily values and computes the monthly averages for each pixel.
}
\examples{
\dontrun{
month_PWC <- daily_to_monthly(input_rack = PWC.stack.y)
month_HS <- daily_to_monthly(input_rack = HS.stack.y)
}
}
