% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAY2ANN.R
\name{DAY2ANN}
\alias{DAY2ANN}
\title{DAY2ANN}
\usage{
DAY2ANN(CROP_IRR, WBGT, ETA, YEAR_INPUT = NULL, ...)
}
\arguments{
\item{CROP_IRR}{choice of crop and irrigation practice}

\item{WBGT}{choice of heat stress function}

\item{ETA}{choice of labor-heat response function}

\item{YEAR_INPUT}{a vector of years of interest}

\item{...}{individual .nc file name of climate variables in order}
}
\value{
A list of two data frames of annual gridded WBGT and PWC
}
\description{
DAY2ANN
}
\examples{
\dontrun{
annual_output <- DAY2ANN(CROP_IRR = "MAIZ_I", WBGT = WBGT, ETA = ETA, YEAR_INPUT = 2027,
"hurs_day_GFDL-ESM2M_rcp60_r1i1p1_EWEMBI_20210101-20301231.nc4",
"tas_day_GFDL-ESM2M_rcp60_r1i1p1_EWEMBI_20210101-20301231.nc4",
"ps_day_GFDL-ESM2M_rcp60_r1i1p1_EWEMBI_20210101-20301231.nc4")
}
}
