% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAY2MON.R
\name{DAY2MON}
\alias{DAY2MON}
\title{DAY2MON}
\usage{
DAY2MON(input_rack)
}
\arguments{
\item{input_rack}{Raster rack of daily value}
}
\value{
A raster rack of monthly value
}
\description{
DAY2MON
}
\examples{
\dontrun{
month_PWC <- DAY2MON(input_rack = PWC.stack.y)
month_HS <- DAY2MON(input_rack = HS.stack.y)
}
}
