% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/G2R.R
\name{G2R}
\alias{G2R}
\title{G2R}
\usage{
G2R(grid_annual_value, SECTOR, rast_boundary)
}
\arguments{
\item{grid_annual_value}{data.frame of grid-level annual value}

\item{SECTOR}{choice of sector}

\item{rast_boundary}{rasterized regional boundaries, 0.5 degree, Set to global extent}
}
\value{
A sf data frame of regional annual value
}
\description{
G2R
}
\examples{
\dontrun{
reg_annual_value <- G2R(grid_annual_value = pwc.hothaps.ann, SECTOR = "MAIZ_R", rast_boundary = country_raster)
}
}
