% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_example_data.R
\name{get_example_data}
\alias{get_example_data}
\title{get_example_data}
\usage{
get_example_data(
  write_dir = getwd(),
  dir_name = "HELPS_Example_Data",
  data_links =
    c("https://zenodo.org/records/14270969/files/monthly_inputs_example.zip?download=1",
    "https://zenodo.org/records/14270969/files/daily_inputs_example.zip?download=1")
)
}
\arguments{
\item{write_dir}{Default = getwd()}

\item{dir_name}{= "HELPS_Example_Data". Name of directory to install zip file to.}

\item{data_link}{Default =
c(""https://zenodo.org/records/14270969/files/daily_inputs_example.zip?download=1",
"https://zenodo.org/records/14270969/files/monthly_inputs_example.zip?download=1")}
}
\value{
path with example input data
}
\description{
Function to download the test data files from zenodo that are needed to run HELPS.
}
\examples{
\dontrun{
library(HELPS)
HELPS::get_example_data()
}
}
\keyword{test}
