% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeatStress.R
\name{HeatStress}
\alias{HeatStress}
\title{HeatStress}
\usage{
HeatStress(SECTOR, HS, YEAR_INPUT, ...)
}
\arguments{
\item{SECTOR}{choice of sector}

\item{HS}{choice of heat stress function}

\item{YEAR_INPUT}{a vector of years of interest}

\item{...}{individual .nc file name of climate variables in order}
}
\value{
A raster stack of daily grid-level heat stress level
}
\description{
HeatStress
}
\examples{
\dontrun{
GD_HS <- DAY2ANN(SECTOR = "MAIZ_I", HS = WBGT1, YEAR_INPUT = 2027, TempRes = "Day",
"hurs_day_GFDL-ESM2M_rcp60_r1i1p1_EWEMBI_20210101-20301231.nc4",
"tas_day_GFDL-ESM2M_rcp60_r1i1p1_EWEMBI_20210101-20301231.nc4",
"ps_day_GFDL-ESM2M_rcp60_r1i1p1_EWEMBI_20210101-20301231.nc4")
}
}
