% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LHR_Foster.R
\name{LHR_Foster}
\alias{LHR_Foster}
\title{LHR_Foster}
\usage{
LHR_Foster(WBGT, workload = NULL)
}
\arguments{
\item{WBGT}{web bulb globe temperature (C)}

\item{workload}{not required in this LHR, but keep it to be consistent with other LHRs
This function is from Foster et al (2021) Table 3
An advanced empirical model for quantifying the impact of heat and climate change on human physical work capacity
adjustment in this package: PWC = 1 for negative WBGT
For additional scientific adjustments, see Smallcombe et al. (2022) Table 2
https://link.springer.com/article/10.1007/s00484-022-02370-7}
}
\value{
value of physical work capacity, range from 0 to 1
}
\description{
LHR_Foster
}
\examples{
\dontrun{
PWC.Foster.Ag <- LHR_Foster(WBGT = 27)
}
}
