% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LHR_Hothaps.R
\name{LHR_Hothaps}
\alias{LHR_Hothaps}
\title{LHR_Hothaps}
\usage{
LHR_Hothaps(WBGT, workload)
}
\arguments{
\item{WBGT}{web bulb globe temperature (C)}

\item{workload}{'high', 'moderate', 'low'
This function is from the high occupational temperature health and productivity suppression (Hothaps) programme
Derived by Kjellstrom et al. (2018) adopted by Bröde et al. (2018) and Orlov et al. (2020).}
}
\value{
value of physical work capacity, range from 0 to 1
}
\description{
LHR_Hothaps
}
\examples{
\dontrun{
PWC.Hothaps.Ag <- LHR_Hothaps(WBGT = 27)
}
}
