% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LHR_ISO.R
\name{LHR_ISO}
\alias{LHR_ISO}
\title{LHR_ISO}
\usage{
LHR_ISO(WBGT, workload)
}
\arguments{
\item{WBGT}{relative humidity (\%)}

\item{workload}{'high', 'moderate', 'low', specifying the rate of metabolic heat production (Watts)
This function is from Kjellstrom et al. (2014)
Based on Kjellstrom et al (2009):
M = 200W for office desk work and service industries
M = 300W for average manufacturing industry work
M = 400W for construction or agricultural work
M = 117W for resting}
}
\value{
value of physical work capacity, range from 0 to 1
}
\description{
LHR_ISO
}
\examples{
\dontrun{
PWC.ISO.Ag <- LHR_ISO(WBGT = 27, workload = 'high')
}
}
