% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LHR_NIOSH.R
\name{LHR_NIOSH}
\alias{LHR_NIOSH}
\title{LHR_NIOSH}
\usage{
LHR_NIOSH(WBGT, workload)
}
\arguments{
\item{WBGT}{relative humidity (\%)}

\item{workload}{'high', 'moderate', 'low', specifying the rate of metabolic heat production (Watts)
This function is from Bröde et al (2018)
Choose National Institute for Occupational Safety and Health (NIOSH) standard; (Jacklitsch et al. 2016)
Based on Kjellstrom et al (2009):
M = 200W for office desk work and service industries
M = 300W for average manufacturing industry work
M = 400W for construction or agricultural work
M = 117W for resting}
}
\value{
value of physical work capacity, range from 0 to 1
}
\description{
LHR_NIOSH
}
\examples{
\dontrun{
PWC.NIOSH.Ag <- LHR_NIOSH(WBGT = 27, M = 400)
}
}
