% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MON2ANN.R
\name{MON2ANN}
\alias{MON2ANN}
\title{MON2ANN}
\usage{
MON2ANN(input_rack, SECTOR)
}
\arguments{
\item{input_rack}{Raster rack of monthly value}

\item{SECTOR}{choice of sector}
}
\value{
A data frame of grid-level annual value
}
\description{
MON2ANN
}
\examples{
\dontrun{
annual_PWC <- MON2ANN(input_rack = month_PWC, SECTOR = "MAIZ_I")
annual_HS <- MON2ANN(input_rack = month_HS, SECTOR = "MAIZ_I")
}
}
