% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly_to_annual.R
\name{monthly_to_annual}
\alias{monthly_to_annual}
\title{monthly_to_annual}
\usage{
monthly_to_annual(input_rack, SECTOR)
}
\arguments{
\item{input_rack}{Raster rack of monthly value}

\item{SECTOR}{choice of sector}
}
\value{
A data frame of grid-level annual value
}
\description{
Calculate Annual Grid-Level Values from Monthly Raster Data
The \code{monthly_to_annual} function computes annual grid-level values from a raster stack of monthly data,
This function supports a wide range of agricultural sectors and accounts for variations in crop calendars.
}
\examples{
\dontrun{
annual_PWC <- monthly_to_annual(input_rack = month_PWC, SECTOR = "MAIZ_I")
annual_HS <- monthly_to_annual(input_rack = month_HS, SECTOR = "MAIZ_I")
}
}
