% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WBGT_ESI.R
\name{WBGT_ESI}
\alias{WBGT_ESI}
\title{WBGT_ESI}
\usage{
WBGT_ESI(hurs, tas, rsds)
}
\arguments{
\item{hurs}{relative humidity (\%)}

\item{tas}{2m air temperature (K)}

\item{sdsr}{surface downward solar radiation (W m-2)}

\item{...}{individual .nc file name of climate variables in order
Environmental Stress Index (ESI) is from Moran et al (2001) & Moran et al (2003)}
}
\value{
A raster stack of daily grid-level heat stress level
}
\description{
WBGT_ESI
}
\examples{
\dontrun{
ESI.test <- WBGT_ESI(50, 303.15, 200)
}
}
