% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WBGT_sun.R
\name{WBGT_sun}
\alias{WBGT_sun}
\title{WBGT_sun}
\usage{
WBGT_sun(hurs, tas, ps)
}
\arguments{
\item{hurs}{relative humidity (\%)}

\item{tas}{2m air temperature (K)}

\item{ps}{air pressure (Pa)}

\item{...}{individual .nc file name of climate variables in order
This function is from Dunne et al (2013)}
}
\value{
A raster stack of daily grid-level heat stress level
}
\description{
WBGT_sun
}
\examples{
\dontrun{
WBGT.sun <- WBGT_sun(50, 303.15, 100000)
expect WBT = 22.25
expect WBGT.shade = 24.57
}
}
