% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LHR_Hothaps.R
\name{LHR_Hothaps}
\alias{LHR_Hothaps}
\title{LHR_Hothaps}
\usage{
LHR_Hothaps(WBGT, workload)
}
\arguments{
\item{WBGT}{web bulb globe temperature (C)}

\item{workload}{'high', 'moderate', 'low'
This function is from Foster et al (2021)
For additional adjustment, see Smallcombe et al. (2022)}
}
\value{
value of physical work capacity, range from 0 to 1
}
\description{
LHR_Hothaps
}
\examples{
\dontrun{
PWC.Hothaps.Ag <- LHR_Hothaps(WBGT = 27)
}
}
